# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OC-Facturer API
  description: API de gestion de la facturation des bons de commande Sewan (OC-DB + Sophia GraphQL)
  version: 0.1.0
  contact:
    name: Dominique Pineau
    email: dominique.pineau@sewan.fr
servers:
  - url: http://localhost:3000
    description: Developpement local
  - url: https://oc-facturer.sewan.fr
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (1h). Obtenu via POST /api/auth/login.
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Erreur serveur
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          description: Nombre total d'elements
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@sewan.fr
        password:
          type: string
          format: password
          example: Admin123!
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT access token (expire 1h)
        refreshToken:
          type: string
          description: JWT refresh token (expire 7j)
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            nom:
              type: string
            email:
              type: string
            accesAdv:
              type: boolean
            accesAdmin:
              type: boolean
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    Utilisateur:
      type: object
      properties:
        id_utilisateur:
          type: string
          format: uuid
        nom_utilisateur:
          type: string
        email_utilisateur:
          type: string
        last_login:
          type: string
          format: date-time
          nullable: true
        acces_commerce:
          type: integer
        acces_adv:
          type: integer
        acces_planif:
          type: integer
        acces_admin:
          type: integer
        acces_cdp:
          type: integer
        acces_attribution:
          type: integer
        acces_deploiement:
          type: integer
        acces_rgpd:
          type: integer
        actif:
          type: integer
    CreateUserRequest:
      type: object
      required: [nom, email, password]
      properties:
        nom:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: "Min 8 chars, 1 majuscule, 1 minuscule, 1 chiffre"
        accesAdv:
          type: boolean
        accesAdmin:
          type: boolean
    BonDeCommande:
      type: object
      properties:
        id_bon_de_commande:
          type: string
          format: uuid
        numero_bdc:
          type: string
        per_name:
          type: string
          description: Nom du client
        dynamics_account_id:
          type: string
          nullable: true
        statut_bdc:
          type: integer
        cree_le:
          type: string
          format: date-time
        commercial_nom:
          type: string
          nullable: true
        type_de_vente:
          type: string
          nullable: true
        ajout_gdc:
          type: integer
          description: "Flag ajout GDC (0/1)"
        gdc_contractId:
          type: string
          format: uuid
          nullable: true
        gdc_contractName:
          type: string
          nullable: true
        gdc_invoicedEntityId:
          type: string
          format: uuid
          nullable: true
        gdc_invoicedEntityName:
          type: string
          nullable: true
    BdcUpdateRequest:
      type: object
      description: "Champs autorises (whitelist): gdc_contractId, gdc_contractName, gdc_invoicedEntityId, gdc_invoicedEntityName, id_sophia_go_facturation, nom_sophia_facturation, ajout_gdc"
      properties:
        gdc_contractId:
          type: string
          format: uuid
        gdc_contractName:
          type: string
        gdc_invoicedEntityId:
          type: string
          format: uuid
        gdc_invoicedEntityName:
          type: string
        ajout_gdc:
          type: integer
    RessourceDpl:
      type: object
      properties:
        id_dpl:
          type: string
          format: uuid
        id_commande:
          type: string
          format: uuid
        code_produit:
          type: string
          nullable: true
        nom:
          type: string
          nullable: true
        quantite:
          type: integer
        nom_site:
          type: string
          nullable: true
        id_site_sophia_go:
          type: string
          format: uuid
          nullable: true
        gdc_contractId:
          type: string
          nullable: true
        gdc_serviceId:
          type: string
          nullable: true
        gdc_categoryId:
          type: string
          nullable: true
        gdc_catalogRef:
          type: string
          nullable: true
        gdc_productName:
          type: string
          nullable: true
        gdc_itemStatus:
          type: string
          nullable: true
          enum: [null, INPROGRESS, ACTIVATED]
        gdc_hidden:
          type: integer
        J2_date_CMES:
          type: string
          format: date
          nullable: true
        date_factu_anticipee:
          type: string
          format: date
          nullable: true
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Connexion utilisateur
      description: Authentification par email/password. Rate-limited (5 tentatives / 15 min par email).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion reussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Email et mot de passe requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Identifiants incorrects
        '429':
          description: Trop de tentatives
          headers:
            Retry-After:
              schema:
                type: integer
              description: Secondes avant prochain essai
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Renouvellement du token
      description: Echange un refresh token valide contre un nouveau couple access + refresh.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens renouveles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Refresh token invalide ou expire
...
  /api/auth/users:
    get:
      tags: [Auth]
      summary: Liste des utilisateurs
      description: Requiert le role admin.
      responses:
        '200':
          description: Liste des utilisateurs (sans mot de passe)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Utilisateur'
        '403':
          description: Acces reserve aux administrateurs
    post:
      tags: [Auth]
      summary: Creer un utilisateur
      description: Requiert le role admin. Mot de passe min 8 chars, 1 majuscule, 1 minuscule, 1 chiffre.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Utilisateur cree
        '400':
          description: Donnees invalides
        '403':
          description: Acces reserve aux administrateurs
        '409':
          description: Email deja utilise
  /api/auth/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Auth]
      summary: Modifier un utilisateur
      description: "Requiert le role admin. Champs modifiables: actif, accesAdv, accesAdmin, password."
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actif:
                  type: boolean
                accesAdv:
                  type: boolean
                accesAdmin:
                  type: boolean
                password:
                  type: string
      responses:
        '200':
          description: Utilisateur mis a jour
        '403':
          description: Acces reserve aux administrateurs
        '404':
          description: Utilisateur non trouve
    delete:
      tags: [Auth]
      summary: Supprimer un utilisateur
      description: Requiert le role admin.
      responses:
        '200':
          description: Utilisateur supprime
        '403':
          description: Acces reserve aux administrateurs

  # ==================== BDC ====================
...
  /api/bdc/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [BDC]
      summary: Detail d'un BDC
      responses:
        '200':
          description: BDC trouve
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/BonDeCommande'
        '404':
          description: BDC non trouve
    patch:
      tags: [BDC]
      summary: Mettre a jour un BDC
      description: Champs autorises (whitelist) uniquement. Requiert auth.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BdcUpdateRequest'
      responses:
        '200':
          description: BDC mis a jour
...
  /api/bdc/facturation:
    get:
      tags: [Facturation]
      summary: Ressources a facturer
      description: Liste paginee avec onglets CMES / Facturations Anticipees. Jointures sur BDC, site, client, duree.
      parameters:
        - name: tab
          in: query
          required: true
          schema:
            type: string
            enum: [cmes, fac_anticipees]
        - name: filter
          in: query
          schema:
            type: string
            enum: [tous, a_facturer, masquees, dans_gdc]
            default: tous
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: sortKey
          in: query
          schema:
            type: string
            enum: [numero_bdc, per_name, code_produit, duree_mois, gdc_productName, quantite, abo, fas, J2_date_CMES, date_factu_anticipee, nom_site, gdc_itemStatus]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Liste paginee
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RessourceDpl'
...